+++ ayaka/config.py
@@ -6,7 +6,7 @@ from typing import Optional
 from pydantic import ValidationError, BaseModel, validator
 from ayaka_utils import ensure_dir_exists
 
-AYAKA_VERSION = "0.0.4.3b0"
+AYAKA_VERSION = "0.0.4.3"
 
 
 class AyakaConfig(BaseModel):