+++ website/change-log.md
@@ -120,3 +120,7 @@ base_send_group，base_send_private，base_send_group_many
 ### 0.0.4.2
 
 - 优化代码结构
+- 变更：cat.on_cmd取消cmds参数的缺省值
+- 变更：console、gocq适配器的run方法，新增自定义参数
+- 新特性：适配器增加on_shutdown钩子
+-